Move admin field helpers to util/adminFields
The CSV export route needs responseFields and formatValue server-side, and
the module holds no Tailwind classes, so the @source rule that kept it
under components/ does not apply. Moving it avoids a util/ -> components/
import from the new server module.

Pure rename plus the two import sites.
